# backend/openapi/schema.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Podoria CRM API",
        "version": "1.0.0",
        "description": "Versioned API contract for Podoria CRM."
    },
    "paths": {
        "/api/v1/audit-events": {
            "get": {
                "operationId": "audit_event_list",
                "summary": "List append-only audit events for an administrator",
                "parameters": [
                    {
                        "in": "query",
                        "name": "actor_id",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "in": "query",
                        "name": "date_from",
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "in": "query",
                        "name": "date_to",
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "in": "query",
                        "name": "section",
                        "schema": {
                            "enum": [
                                "accounts",
                                "team",
                                "settings",
                                "patients",
                                "scheduling",
                                "medical",
                                "visits",
                                "billing",
                                "cash",
                                "inventory"
                            ],
                            "type": "string",
                            "minLength": 1
                        },
                        "description": "* `accounts` - accounts\n* `team` - team\n* `settings` - settings\n* `patients` - patients\n* `scheduling` - scheduling\n* `medical` - medical\n* `visits` - visits\n* `billing` - billing\n* `cash` - cash\n* `inventory` - inventory"
                    }
                ],
                "tags": [
                    "audit"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuditEventListResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/audit-events/{event_id}": {
            "get": {
                "operationId": "audit_event_retrieve",
                "summary": "Return redacted before/after details for one audit event",
                "parameters": [
                    {
                        "in": "path",
                        "name": "event_id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "audit"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuditEventDetail"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuditActor": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "email": {
                        "oneOf": [
                            {
                                "type": "string",
                                "format": "email"
                            },
                            {
                                "type": "string",
                                "maxLength": 0
                            }
                        ]
                    },
                    "role": {
                        "type": "string"
                    }
                },
                "required": [
                    "display_name",
                    "email",
                    "id",
                    "role"
                ]
            },
            "AuditChange": {
                "type": "object",
                "properties": {
                    "field": {
                        "type": "string"
                    },
                    "before": {
                        "nullable": true
                    },
                    "after": {
                        "nullable": true
                    }
                },
                "required": [
                    "after",
                    "before",
                    "field"
                ]
            },
            "AuditEventDetail": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "occurred_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "actor": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AuditActor"
                            }
                        ],
                        "readOnly": true
                    },
                    "section": {
                        "type": "string",
                        "maxLength": 32
                    },
                    "action": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "object": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AuditObject"
                            }
                        ],
                        "readOnly": true
                    },
                    "result": {
                        "type": "string",
                        "maxLength": 32
                    },
                    "description": {
                        "type": "string"
                    },
                    "before": {},
                    "after": {},
                    "changes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuditChange"
                        },
                        "readOnly": true
                    },
                    "note": {
                        "type": "string"
                    },
                    "correlation_id": {
                        "type": "string",
                        "maxLength": 128
                    }
                },
                "required": [
                    "action",
                    "actor",
                    "changes",
                    "correlation_id",
                    "id",
                    "object",
                    "occurred_at",
                    "section"
                ]
            },
            "AuditEventListItem": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "occurred_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "actor": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AuditActor"
                            }
                        ],
                        "readOnly": true
                    },
                    "section": {
                        "type": "string",
                        "maxLength": 32
                    },
                    "action": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "object": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AuditObject"
                            }
                        ],
                        "readOnly": true
                    },
                    "result": {
                        "type": "string",
                        "maxLength": 32
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "action",
                    "actor",
                    "id",
                    "object",
                    "occurred_at",
                    "section"
                ]
            },
            "AuditEventListResponse": {
                "type": "object",
                "properties": {
                    "events": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuditEventListItem"
                        }
                    },
                    "next_cursor": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    }
                },
                "required": [
                    "events",
                    "next_cursor"
                ]
            },
            "AuditObject": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "label",
                    "type"
                ]
            },
...
            "ErrorEnvelope": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "fields": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "correlation_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "correlation_id",
                    "fields",
                    "message"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "podoria_sessionid"
            }
        }
    }
}